requirements.txt
dgl_cuda11.6==0.9.1
helper==2.5.0
networkx==2.8.5
numpy==1.23.1
pandas==1.4.3
rpy2==3.5.1
scikit_learn==1.1.2
scipy==1.9.0
torch==1.12.1
torch_geometric==2.0.4
torch_scatter==2.0.9
